df: move BlockSize to a new module, blocks.rs
